Refactor windowRep code from BLANT into a separate file
Code related to windowRep has been moved from the main blant.c file to a separate file called blant_window.c. As such, some static variables have been made extern. Makefile (specifically make blant) has been updated to reflect this change.
